+++ src/consumer/component/consumer-component.js
@@ -43,6 +43,7 @@ import {
   DEFAULT_REGISTRY_DOMAIN_PREFIX,
   CFG_REGISTRY_DOMAIN_PREFIX
 } from '../../constants';
+import ComponentWithDependencies from '../../scope/component-dependencies';
 
 export type RelativePath = {
   sourceRelativePath: string,
@@ -98,6 +99,7 @@ export default class Component {
   writtenPath: ?string; // needed for generate links
   dependenciesSavedAsComponents: ?boolean = true; // otherwise they're saved as npm packages
   originallySharedDir: ?string; // needed to reduce a potentially long path that was used by the author
+  _wasOriginallySharedDirStripped: ?boolean; // whether stripOriginallySharedDir() method had been called, we don't want to strip it twice
   _writeDistsFiles: ?boolean = true;
   _areDistsInsideComponentDir: ?boolean = true;
   isolatedEnvironment: IsolatedEnvironment;
@@ -466,6 +468,7 @@ export default class Component {
    * imports it. NESTED and AUTHORED components are written as is.
    */
   stripOriginallySharedDir(bitMap: BitMap): void {
+    if (this._wasOriginallySharedDirStripped) return;
     this.setOriginallySharedDir();
     const originallySharedDir = this.originallySharedDir;
     const pathWithoutSharedDir = (pathStr, sharedDir, isLinuxFormat) => {
@@ -500,6 +503,7 @@ export default class Component {
         }
       });
     });
+    this._wasOriginallySharedDirStripped = true;
   }
 
   updateDistsPerConsumerBitJson(consumer: Consumer, componentMap: ComponentMap): void {
@@ -595,20 +599,13 @@ export default class Component {
     if (origin === COMPONENT_ORIGINS.IMPORTED && componentMap.origin === COMPONENT_ORIGINS.NESTED) {
       // when a user imports a component that was a dependency before, write the component directly into the components
       // directory for an easy access/change. Then, remove the current record from bit.map and add an updated one.
-      await this._writeToComponentDir({
-        bitDir: calculatedBitDir,
-        withBitJson,
-        withPackageJson,
-        driver,
-        force,
-        writeBitDependencies,
-        dependencies,
-        deleteBitDirContent
-      });
-      // todo: remove from the file system
+      const oldLocation = path.join(consumerPath, componentMap.rootDir);
+      logger.debug(
+        `deleting the old directory of a component at ${oldLocation}, the new directory is ${calculatedBitDir}`
+      );
+      fs.removeSync(oldLocation);
       bitMap.removeComponent(this.id.toString());
-      this._addComponentToBitMap(bitMap, calculatedBitDir, origin, parent);
-      return this;
+      componentMap = this._addComponentToBitMap(bitMap, calculatedBitDir, origin, parent);
     }
     logger.debug('component is in bit.map, write the files according to bit.map');
     this.updateDistsPerConsumerBitJson(consumer, componentMap);
@@ -966,7 +963,7 @@ export default class Component {
    * find a shared directory among the files of the main component and its dependencies
    */
   setOriginallySharedDir() {
-    if (this.originallySharedDir) return;
+    if (this.originallySharedDir !== undefined) return;
     // taken from https://stackoverflow.com/questions/1916218/find-the-longest-common-starting-substring-in-a-set-of-strings
     // It sorts the array, and then looks just at the first and last items
     const sharedStartOfArray = (array) => {
@@ -987,6 +984,21 @@ export default class Component {
     if (!sharedStart || !sharedStart.includes(pathSep)) return;
     const lastPathSeparator = sharedStart.lastIndexOf(pathSep);
     this.originallySharedDir = sharedStart.substring(0, lastPathSeparator);
+  }
+
+  async toComponentWithDependencies(bitMap: BitMap, consumer: Consumer): Promise<ComponentWithDependencies> {
+    const getDependencies = () => {
+      return this.dependencies.map((dependency) => {
+        if (bitMap.isExistWithSameVersion(dependency.id)) {
+          return consumer.loadComponent(dependency.id);
+        }
+        // when dependencies are imported as npm packages, they are not in bit.map
+        this.dependenciesSavedAsComponents = false;
+        return consumer.scope.loadComponent(dependency.id, false);
+      });
+    };
+    const dependencies = await Promise.all(getDependencies());
+    return new ComponentWithDependencies({ component: this, dependencies });
   }
 
   static fromObject(object: Object): Component {