+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- when a nested dependency is imported directly, re-link all its dependents
+
 ## [0.12.0-ext.11] - 2018-01-02
 
 - move `save-dependencies-as-components` flag from `bit import` command to be configurable in consumer bit.json